用户注册成功发送短信  客服问题为空提交修复  模板消息价格修复

+++ app/models/order.rb
@@ -320,7 +320,7 @@ class Order < ApplicationRecord
         color: "#173177"
       },
       keyword2: {
-        value: price.to_i.to_s + "元",
+        value: price.to_f.to_s + "元",
         color:"#173177"
       },
       keyword3: {
